refactor(agent): split main.rs + tray.rs + response_action_commands.rs
Splits three medium god files in `apps/agent/src-tauri/src/` into focused
submodules so each production file stays under 600 lines:

- `main.rs` (1,386 -> 332): keeps `fn main`, tauri builder, `AppState`
  and `ShutdownComplete`. `run_agent` + heartbeat loops move out.
- `bootstrap/`: orchestrator. `mod.rs` re-exports `run_agent` plus the
  shared NATS auth-failure / security-config helpers. `run.rs` is the
  slim `run_agent` body. `nats.rs` hosts the enterprise NATS bootstrap
  (policy/telemetry/posture subscribers + approval outbox). `event_loops`
  hosts spawn helpers for daemon-state/policy/SSE/approval consumers.
  `listeners.rs` registers the four tauri front-end listeners.
- `tasks/heartbeats.rs`: NATS + local heartbeat loops, lifted whole.
- `tray/` directory: `mod.rs` is just `setup_tray` + menu IDs. `menu.rs`
  builds the menu tree. `manager.rs` is the `TrayManager` actor +
  `TrayState`. `menu_handlers.rs` routes click events. `dashboard_url.rs`
  is the pure URL validation / resolution helpers (extractable and
  tested in isolation).
- `response_action_commands/` directory: `mod.rs` re-exports the
  handler. `dto.rs` is the wire DTOs + the `PolicyRuleDiffValidationCommand`
  + ack-context types. `validate.rs` is signed-envelope parsing and
  command validation. `policy_rule_diff.rs` executes local validation
  and shapes success/failure payloads (incl. signed failure ack).
  `control_ack.rs` is the Control-API postback config + dispatch +
  retry-sink enqueue. `handler.rs` is the NATS subscriber lifecycle.
  `tests.rs` keeps the existing test suite untouched.

No production file > 600 lines in these three areas. `cargo check
--tests`, `cargo clippy --all-targets -- -D warnings`, and the
response_action_commands / tray / bootstrap test subsets all pass.
